test(zeebe): make versionTag tooltip test work
This test was previously failing, but async, outside
of the test cycle.

It was also broken, because it did not toggle open the
general tab; so hovering over the actual `versionTag`
field could never happen. Now we clearly model the user
journey.

Bonus points for getting rid of `clock`. We can simply
use `waitFor` to wait for a DOM element expected for
arrival.